+++ test/base_repository.test.ts
@@ -1,10 +1,11 @@
 import {expect} from "chai";
 
-import {BaseRepository, entity, id, repository, TYPES} from "../src";
+import {BaseRepository, entity, excludeFromIndex, id, repository, TYPES} from "../src";
 import {Container} from "inversify";
 import {Datastore} from "@google-cloud/datastore";
 import {anything, capture, instance, mock, verify, when} from "ts-mockito";
 import {IsEmail} from "class-validator";
+import {Expose} from "class-transformer";
 
 describe("Unit Test: BaseRepository", () => {
 
@@ -24,6 +25,8 @@ describe("Unit Test: BaseRepository", () => {
     class MyEntity {
       @id()
       public entityId: string;
+      @excludeFromIndex()
+      public unindexedProperty: string;
     }
 
     @repository(MyEntity)
@@ -37,6 +40,7 @@ describe("Unit Test: BaseRepository", () => {
     // Give Parameters
     const t = new MyEntity();
     t.entityId = "foo";
+    t.unindexedProperty = "bar";
 
     let mockKey = {};
 
@@ -53,7 +57,43 @@ describe("Unit Test: BaseRepository", () => {
       expect(result).eql(t);
       expect(saveRequest.key).eql(mockKey);
       expect(saveRequest.data).eql(t);
+      expect(saveRequest.excludeFromIndexes).contains("unindexedProperty");
 
+      done();
+    });
+  });
+
+  it("should save an entity and excludeExtraneousValues when decorated", (done) => {
+
+    @entity("MyEntityKind", {excludeExtraneousValues: true})
+    class MyEntityIgnoringValidation {
+      @id()
+      @Expose()
+      public entityId: string;
+      public transientField: string;
+    }
+
+    @repository(MyEntityIgnoringValidation)
+    class TestRepositoryIgnoringValidation extends BaseRepository<MyEntityIgnoringValidation> {
+    }
+
+    // Setup Container
+    container.bind<TestRepositoryIgnoringValidation>(TestRepositoryIgnoringValidation).toSelf();
+    const fixture = container.get<TestRepositoryIgnoringValidation>(TestRepositoryIgnoringValidation);
+
+    // Give Parameters
+    const t = new MyEntityIgnoringValidation();
+    t.entityId = "some_id";
+    t.transientField = "some_value";
+
+    // @ts-ignore
+    when(mockDb.save(anything())).thenResolve();
+
+    // Then
+    fixture.save(t).then(result => {
+      const [saveRequest] = capture(mockDb.save).last();
+      expect(saveRequest.data.entityId).eql("some_id");
+      expect(saveRequest.data.transientField).eql(undefined);
       done();
     });
   });
@@ -111,10 +151,10 @@ describe("Unit Test: BaseRepository", () => {
     container.bind<TestRepository>(TestRepository).toSelf();
     const fixture = container.get<TestRepository>(TestRepository);
 
-
     let expResponse = [[
       {entityId: "1", nonMappedValue: ""}
     ], {}];
+
     // @ts-ignore
     when(mockDb.runQuery(anything())).thenResolve(expResponse);