+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.2.0
+
+* Update [Wiki](https://github.com/pflima92/google-cloud-datastore-inversify/wiki)
+* Fix filter `Opeartors`
+* Add `excludeFromIndex` decorator to exclude a specific property from indexes
+* Add `entityOptions` for custom configurations
+* Add `excludeExtraneousValues` that allows ensure all properties from the type class
+
 ## v0.1.2
 
 * Export `TYPES` for `Datastore` binding